Issue-115:Sonar Scanner fixes + streets_service namespace
