+++ streets_utils/streets_service_base/src/streets_configuration_exception.cpp
@@ -1,5 +1,8 @@
 #include "streets_configuration_exception.h"
 
-// Constructor
-streets_configuration_exception::streets_configuration_exception(const std::string &msg ): std::runtime_error(msg){};
-streets_configuration_exception::~streets_configuration_exception() = default;
+namespace streets_service {
+    // Constructor
+    streets_configuration_exception::streets_configuration_exception(const std::string &msg ): std::runtime_error(msg){};
+    // Destructor
+    streets_configuration_exception::~streets_configuration_exception() = default;
+}
